+++ immutable-arrays/transformations-to-standard-collections/src/main/kotlin/com/danrusu/pods4k/immutableArrays/TransformationsToSet.kt
@@ -9,6 +9,7 @@ import kotlin.Float
 import kotlin.Int
 import kotlin.Long
 import kotlin.Short
+import kotlin.collections.MutableSet
 import kotlin.collections.Set
 
 /**
@@ -55,3 +56,48 @@ public fun ImmutableFloatArray.toSet(): Set<Float> = asList().toSet()
  * See [Array.toSet]
  */
 public fun ImmutableDoubleArray.toSet(): Set<Double> = asList().toSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun <T> ImmutableArray<T>.toMutableSet(): MutableSet<T> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableBooleanArray.toMutableSet(): MutableSet<Boolean> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableByteArray.toMutableSet(): MutableSet<Byte> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableCharArray.toMutableSet(): MutableSet<Char> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableShortArray.toMutableSet(): MutableSet<Short> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableIntArray.toMutableSet(): MutableSet<Int> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableLongArray.toMutableSet(): MutableSet<Long> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableFloatArray.toMutableSet(): MutableSet<Float> = asList().toMutableSet()
+
+/**
+ * See [Array.toMutableSet]
+ */
+public fun ImmutableDoubleArray.toMutableSet(): MutableSet<Double> = asList().toMutableSet()