+++ immutable-arrays/transformations-to-standard-collections/src/test/kotlin/com/danrusu/pods4k/immutableArrays/TransformationsToSetTest.kt
@@ -2,11 +2,13 @@ package com.danrusu.pods4k.immutableArrays
 
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
+import strikt.assertions.isA
+import strikt.assertions.isEmpty
 import strikt.assertions.isEqualTo
 
 class TransformationsToSetTest {
     @Test
-    fun `toList validation`() {
+    fun `toSet validation`() {
         // empty
         with(emptyImmutableArray<String>()) {
             expectThat(toSet()).isEqualTo(emptySet())
@@ -22,6 +24,30 @@ class TransformationsToSetTest {
         with(immutableArrayOf(7, 15, 15)) {
             expectThat(toSet())
                 .isEqualTo(setOf(7, 15))
+        }
+    }
+
+    @Test
+    fun `toMutableSet validation`() {
+        // empty
+        with(emptyImmutableArray<String>()) {
+            expectThat(toMutableSet())
+                .isA<LinkedHashSet<String>>()
+                .isEmpty()
+        }
+        with(emptyImmutableIntArray()) {
+            expectThat(toMutableSet())
+                .isA<LinkedHashSet<String>>()
+                .isEmpty()
+        }
+        // multiple elements
+        with(immutableArrayOf("Bob", "Jill", "Bob")) {
+            expectThat(toMutableSet())
+                .isEqualTo(mutableSetOf("Bob", "Jill"))
+        }
+        with(immutableArrayOf(7, 15, 15)) {
+            expectThat(toMutableSet())
+                .isEqualTo(mutableSetOf(7, 15))
         }
     }
 }